requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-telegram-bot==21.3
psycopg2-binary==2.9.9
apscheduler==3.10.4
httpx==0.27.0
